draft ppe:refactor
